Split datasets into separate files
